refactor clean ast out of MainAir
